refactor: assessment and evaluation logic
- Refactor AssessmentController and AssessmentStudentController
- Update DTOs for assessment and evaluation data
- Modify AssessmentAnswer and AssessmentQuestion models
- Update AssessmentStudentRepository with new queries
- Refactor AssessmentService and AssessmentStudentService
- Change score columns to DECIMAL(5,2) in database migration

+++ src/main/java/com/use3w/grade/dto/AssessmentDetailsDTO.java
@@ -9,6 +9,6 @@ public record AssessmentDetailsDTO(
         Set<AssessmentDetailsClassDTO> classes
 ) {
     public record AssessmentDetailsClassDTO(
-            UUID id, String name, Long countEvaluatedStudents, Integer countStudents
+            UUID id, String name, Integer countEvaluatedStudents, Integer countStudents
     ) {}
 }
